+++ openilc/__init__.py
@@ -1,11 +1,17 @@
 from .hilc import HILC
 from .nilc import NILC
 from .configs import ConfigTable, load_csv_table, load_table
+from .sht import DuccAdjointSHT, DuccPseudoSHT, DuccSHT, HealpySHT, get_sht_backend
 
 __all__ = [
     "ConfigTable",
+    "DuccAdjointSHT",
+    "DuccPseudoSHT",
+    "DuccSHT",
     "HILC",
+    "HealpySHT",
     "NILC",
+    "get_sht_backend",
     "load_csv_table",
     "load_table",
 ]